Move toolbar menu to own component
