+++ client/src/queryEditor/toolbar/Toolbar.js
@@ -1,23 +1,16 @@
-import { MenuItem } from '@reach/menu-button';
 import CopyIcon from 'mdi-react/ContentCopyIcon';
 import UnsavedIcon from 'mdi-react/ContentSaveEditIcon';
 import SaveIcon from 'mdi-react/ContentSaveIcon';
 import DatabaseIcon from 'mdi-react/DatabaseIcon';
-import DotsVerticalIcon from 'mdi-react/DotsVerticalIcon';
 import FormatIcon from 'mdi-react/FormatAlignLeftIcon';
 import NewIcon from 'mdi-react/PlusIcon';
 import TagsIcon from 'mdi-react/TagMultipleIcon';
 import PropTypes from 'prop-types';
 import React, { useState } from 'react';
-import { Redirect } from 'react-router-dom';
 import { connect } from 'unistore/react';
 import Button from '../../common/Button';
-import Drawer from '../../common/Drawer';
 import IconButton from '../../common/IconButton';
-import IconMenu from '../../common/IconMenu';
 import Input from '../../common/Input';
-import ConnectionListDrawer from '../../connections/ConnectionListDrawer';
-import ConnectionAccessListDrawer from '../../connectionAccesses/ConnectionAccessListDrawer';
 import {
   formatQuery,
   handleCloneClick,
@@ -27,19 +20,15 @@ import {
   setQueryState
 } from '../../stores/queries';
 import { toggleSchema } from '../../stores/schema';
-import UserList from '../../users/UserList';
-import fetchJson from '../../utilities/fetch-json.js';
 import ConnectionDropDown from '../ConnectionDropdown';
-import QueryHistoryModal from '../../queryHistory/QueryHistoryModal';
-import AboutModal from './AboutModal';
 import ChartButton from './ChartButton';
 import QueryListButton from './QueryListButton';
 import QueryTagsModal from './QueryTagsModal';
+import ToolbarMenu from './ToolbarMenu';
 import ToolbarSpacer from './ToolbarSpacer';
 
 function mapStateToProps(state) {
   return {
-    currentUser: state.currentUser,
     isRunning: state.isRunning,
     isSaving: state.isSaving,
     queryId: state.query && state.query._id,
@@ -60,7 +49,6 @@ const ConnectedEditorNavBar = connect(mapStateToProps, store => ({
 }))(React.memo(Toolbar));
 
 function Toolbar({
-  currentUser,
   formatQuery,
   handleCloneClick,
   isRunning,
@@ -76,42 +64,9 @@ function Toolbar({
   unsavedChanges
 }) {
   const [showTags, setShowTags] = useState(false);
-  const [showUsers, setShowUsers] = useState(false);
-  const [redirectToSignIn, setRedirectToSignIn] = useState(false);
-  const [showQueryHistory, setShowQueryHistory] = useState(false);
-  const [showAbout, setShowAbout] = useState(false);
-  const [showConnections, setShowConnections] = useState(false);
-  const [showConnectionAccesses, setShowConnectionAccesses] = useState(false);
 
   const error = showValidation && !queryName.length;
   const cloneDisabled = !queryId;
-
-  const isAdmin = currentUser.role === 'admin';
-
-  if (redirectToSignIn) {
-    return <Redirect push to="/signin" />;
-  }
-
-  // Reach UI menu / IconMenu does not like the {someBoolean && <element/>} patterm
-  // As a workaround optional MenuItems are managed in an array
-  let menuItems = [];
-  if (isAdmin) {
-    menuItems = [
-      <MenuItem key="connections" onSelect={() => setShowConnections(true)}>
-        Connections
-      </MenuItem>,
-      <MenuItem key="users" onSelect={() => setShowUsers(true)}>
-        Users
-      </MenuItem>,
-      <MenuItem
-        key="connectionAccessess"
-        style={{ borderBottom: '1px solid #ddd' }}
-        onSelect={() => setShowConnectionAccesses(true)}
-      >
-        Connection Accesses
-      </MenuItem>
-    ];
-  }
 
   return (
     <div
@@ -191,57 +146,7 @@ function Toolbar({
 
         <ToolbarSpacer grow />
 
-        <IconMenu icon={<DotsVerticalIcon aria-label="menu" />}>
-          {menuItems}
-          <MenuItem
-            key="queryHistory"
-            style={{ borderBottom: '1px solid #ddd' }}
-            onSelect={() => setShowQueryHistory(true)}
-          >
-            Query History
-          </MenuItem>
-          <MenuItem
-            style={{ borderBottom: '1px solid #ddd' }}
-            onSelect={() => setShowAbout(true)}
-          >
-            About
-          </MenuItem>
-          <MenuItem
-            onSelect={async () => {
-              await fetchJson('GET', '/api/signout');
-              setRedirectToSignIn(true);
-            }}
-          >
-            Sign out
-          </MenuItem>
-        </IconMenu>
-
-        <Drawer
-          title={'Users'}
-          visible={showUsers}
-          width={600}
-          onClose={() => setShowUsers(false)}
-          placement={'right'}
-        >
-          <UserList />
-        </Drawer>
-
-        <QueryHistoryModal
-          visible={showQueryHistory}
-          onClose={() => setShowQueryHistory(false)}
-        />
-
-        <AboutModal visible={showAbout} onClose={() => setShowAbout(false)} />
-
-        <ConnectionListDrawer
-          visible={showConnections}
-          onClose={() => setShowConnections(false)}
-        />
-
-        <ConnectionAccessListDrawer
-          visible={showConnectionAccesses}
-          onClose={() => setShowConnectionAccesses(false)}
-        />
+        <ToolbarMenu />
       </div>
     </div>
   );
